Bringing back old mult-threaded solution as an option.
Turns out it was never wrong (I thought it was due to sleep deprivation). The only thing with the new multithreaded binary splitting function is that a barrier is used to make sure all worker threads are spawned before any of the others start taking CPU time (since worker threads have higher priority).